WIP on loopy scrolling
This begins to implement true scrolling behavior as it is implemented in
the PPU hardware. Trying to make this a hybrid ppu (emulating hardware
and emulating behavior) doesn't really make sense. For more accurate
scrolling I think the whole ppu needs to run at a cycle level, and the
shift registers need to be implemented for real.