Make the PPU run at 60 fps at all times
Prior to this, we were drawing frames at 60 fps, but the ppu's state machine was not tied to 60 iterations per second. This commit does:
1. Move all gfx interaction out of nes.cpp and into gfx.cpp
2. Draw the frame on every completion of the the full Ppu state machine.

This improves the accuracty of our real time rendering. For example, the transition when walking through the very first door in SMB2 is now correct. Prior to this the transition happened immediately and the sound would jump. This was caused by the Ppu not signalling to draw a frame if rednering was disabled.