Rewrite nes.cpp and restructure things for extensibility
This commit restructures the overall nes object and methods in a way that
should make it easier to write an external game loop around the nes.

1. Add Nes::DoFrame and rewrite the Nes::Run method

I was able to isolate the external calls to check for input and blit the
screen outside of the actual nes frame processing. DoFrame will process
one screens worth of pixel data, so it should be used if writing an
external game loop. Run should be used (with an implementation of IGfx and
IInput) to run the nes by itself, without an external game loop.

2. Continue to isolate all SDL references

SDL will continue to be our defualt implementation for rendering, audio,
and input. If nes is launched from command line on any platform, it should
be using SDL.

The goal is to make the core nes code the most usable by an application
wishing to host the emulator.