Rewrite Input for extensibility.
This change puts the new infrastructure in place. This includes:

1. IControllerPortDevice Interface: Basic interface for any device that
is connected to a controller port and read by the CPU

2. IStandardController Interface: Basic interface representing a standard
controller with 8 pushable buttons

3. StandardController Class: Implementation of the IControllerPortDevice
and IStandardController interfaces that makes a standard controller work.

4. Moves the old IInput interface to IHostInput. This is only needed if
using the Nes::Run function, which takes an IHostInput*. This allows Nes
to check for input from the host. Implementations of IHostInput should
internally drive an IStandardController* that it obtained from the Nes.

TODO:
I still need to figure out the exact lifetime management for input
devices that are created by Nes but handed out to callers.