Implement lifetime management for standard controllers
Come up with a memory management system for controllers. When a controller
is requested, it will created by Nes and a bare pointer to it will be
returned to the caller. If there was an exiting IControllerPortDevice
attached to the device, it's memory will be freed and all pointers
previously handed out will be invalid.