Improve handling of simultaneous left/right up/down
This improves our behavior for handling simultaneous up/down, left/right
presses. If at any point we detect that both up and down or both left and
right are pressed, we treat it as if neither of them are. We also
correctly handle when buttons are released from a both pressed state. This
also introduces a mutex and lock_guard pattern to keep the button presses
thread safe.