Project Restructure and initial UWP App
This commit restructures the project slightly yet again.

1. Move all build files out of src. The top level windows dir contains
all build and source files specific to windows platforms. Nes is now built
as a static library. The motivation for this was so that both the dll
project and the native windows runtime component could simply link against
the same static lib rather than try to include all the same files in their
projects.

2. Initial implementation of Nes as a native windows runtime component.
This provides a simple C++/CX wrapper that allows Nes to be constructed
and used inside of any windows runtime application.

TODO: Improve the asynchronicity of the API.

3. Initial UWP for Nes. The initial UWP is written in C# and references
the native windows runtime component. Today it opens a file picker
immediately when the app starts and will start playing the nes file that
is chosen. It also adds file association for .nes and enough simple code
to make file activation work.

TODO: Lots. This is a pretty barebones app that is the base to continue
building a full featured emulator application from.

4. Add configurations for x86, x64, and ARM. For phone scenarios we need
an ARM build and for UWP I imagine we will want primarily x64. I am fine
if we only support debug support on x86 if that is easier, but I wanted to
be able to at least build everything.

TODO: The Nes static lib cannot be built for ARM because currently it
tries to build as a desktop project. This breaks the ARM runtime
component, meaning there is no build of the UWP for ARM currently. I think
I can edit the lib project by hand enough so that for the arm
configuration it builds a universal windows static lib that will build for
arm and we can link into the runtime component.